Added A SQL Database
Added a Button which stores favorite users and after clicking on favorite you can also unfavorite in the same button, all this being in the InfoUserPage.
On the MainPage, added a button which, on click, displays all the favorited users, added another button which clears the search results and added another button which clears the database of any favorited users.